Cleaning up old menu code
