--- app/src/main/res/menu/menu_conversation.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:title="@string/conversationsSettingsAllMedia"
-        android:id="@+id/menu_view_all_media"
-        android:contentDescription="@string/AccessibilityId_conversationsSettingsAllMedia"/>
-
-    <item
-        android:title="@string/search"
-        android:id="@+id/menu_search"
-        android:contentDescription="@string/AccessibilityId_search"
-        app:actionViewClass="androidx.appcompat.widget.SearchView"
-        app:showAsAction="collapseActionView" />
-
-    <item
-        android:title="@string/conversationsAddToHome"
-        android:id="@+id/menu_add_shortcut" />
-
-</menu>
