--- app/src/main/res/menu/menu_conversation_block.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:title="@string/block"
-        android:contentDescription="@string/AccessibilityId_block"
-        android:id="@+id/menu_block" />
-
-</menu>
