--- app/src/main/res/menu/menu_conversation_copy_account_id.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-        <item
-            android:title="@string/accountIDCopy"
-            android:id="@+id/menu_copy_account_id"
-            android:icon="@drawable/ic_copy"
-            android:contentDescription="@string/AccessibilityId_accountIdCopy" />
-</menu>
