--- app/src/main/res/menu/menu_conversation_expiration.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:title="@string/disappearingMessages"
-        android:id="@+id/menu_expiring_messages"
-        android:contentDescription="@string/AccessibilityId_disappearingMessages" />
-
-</menu>
