--- app/src/main/res/menu/menu_conversation_legacy_group.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/menu_edit_group"
-        android:contentDescription="@string/AccessibilityId_groupEdit"
-        android:title="@string/groupEdit"
-        app:showAsAction="collapseActionView" />
-
-    <item
-        android:id="@+id/menu_leave_group"
-        android:contentDescription="@string/AccessibilityId_groupLeave"
-        android:title="@string/groupLeave"
-        app:showAsAction="collapseActionView"/>
-
-</menu>
