--- app/src/main/res/menu/menu_conversation_muted.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:title="@string/notificationsMuteUnmute"
-        android:id="@+id/menu_unmute_notifications" />
-
-</menu>
