--- app/src/main/res/menu/menu_conversation_notification_settings.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:title="@string/sessionNotifications"
-        android:id="@+id/menu_notification_settings"
-        android:contentDescription="@string/AccessibilityId_sessionNotifications"/>
-</menu>
