--- app/src/main/res/menu/menu_conversation_open_group.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:title="@string/communityUrlCopy"
-        android:id="@+id/menu_copy_open_group_url" />
-
-    <item
-        android:title="@string/membersInvite"
-        android:id="@+id/menu_invite_to_open_group" />
-
-</menu>
