--- app/src/main/res/menu/menu_conversation_unblock.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:title="@string/blockUnblock"
-        android:id="@+id/menu_unblock" />
-
-</menu>
