--- app/src/main/res/menu/menu_conversation_unmuted.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:title="@string/notificationsMute"
-        android:id="@+id/menu_mute_notifications"
-        android:contentDescription="@string/AccessibilityId_notificationsMute" />
-
-</menu>
